Partial ECS mostly done, needs cleanup and visualizer
